+++ src/app/ReactQueryProvider/useCreateQueryClient.ts
@@ -1,8 +1,7 @@
 import { useRef } from "react";
 import { QueryClient, QueryCache, MutationCache } from "@tanstack/react-query";
 import { HttpError } from "@/shared/lib";
-import { useOverlayStore } from "@/shared/store";
-import { useSnackbar } from "@/shared/store/snackbar";
+import { useOverlayStore, useSnackbar } from "@/shared/store";
 import { useRefreshToken } from "./errorHandlers";
 
 export const useCreateQueryClient = () => {
