+++ src/features/auth/ui/userInfoRegistrationForm.tsx
@@ -6,8 +6,7 @@ import {
 } from "@/entities/auth/ui";
 import type { Region } from "@/entities/map/types/server";
 import { useModal } from "@/shared/lib";
-import { useAuthStore } from "@/shared/store/auth";
-import { useSnackbar } from "@/shared/store/snackbar";
+import { useAuthStore, useSnackbar } from "@/shared/store";
 import { Badge } from "@/shared/ui/badge";
 import { Button } from "@/shared/ui/button";
 import { ActionChip } from "@/shared/ui/chip";
