+++ src/features/marking/ui/markingFormModal.tsx
@@ -6,8 +6,7 @@ import {
   MARKING_VISIBILITY_MAP,
   MARKING_VISIBILITY_ENTRIES,
 } from "@/entities/marking/constants";
-import { useAuthStore } from "@/shared/store";
-import { useSnackbar } from "@/shared/store/snackbar";
+import { useAuthStore, useSnackbar } from "@/shared/store";
 import { Badge } from "@/shared/ui/badge";
 import { Button } from "@/shared/ui/button";
 import { MyLocationIcon, PlusIcon } from "@/shared/ui/icon";
