+++ src/pages/setting/page.tsx
@@ -2,8 +2,7 @@ import { Link } from "react-router-dom";
 import { Report, LogoutButton } from "@/features/setting/ui";
 import { ROUTER_PATH } from "@/shared/constants";
 import { withAuth } from "@/shared/lib";
-import { useAuthStore } from "@/shared/store";
-import { useSnackbar } from "@/shared/store/snackbar";
+import { useAuthStore, useSnackbar } from "@/shared/store";
 import { DividerLine } from "@/shared/ui/divider";
 import { ArrowRightIcon } from "@/shared/ui/icon";
 import { BackwardNavigationBar } from "@/shared/ui/navigationBar";
