+++ src/shared/store/index.ts
@@ -1,2 +1,3 @@
 export * from "./auth";
 export * from "./overlay";
+export * from "./snackbar";
